⚡ Switch to pubsub to have it update for everyone
Anyone in a topic will see the post pop up or a room banner change.
Added three things to subscribe to, global, room and topic. Posts are
broadcasted to topics, because everyone viewing the topic wants the
update, and banner is broadcasted to the room. Global is not used yet
but will be for example highlight other rooms with unread posts when
someone posts.

TODO: Unsubscribe from the current room or topic when switching to
another.