✨ 💄 Add a room join button for non-members
When entering a room without being a member, show a join button instead
of the input field for writing in the room. When joining also broadcast
a pubsub message so other people in the room gets an updated members
list in the info section.

Fixed long wrapping of long messages, for some reason break-words
did not work as per documentation so opted for break-all.

Also fixed memebers list not being proper list items and gave them an
indent.